uni_front

+++ client_side/pages/sub_pages/order_view.vue
@@ -14,11 +14,11 @@
             <tr v-for="cart_item in cart_items" :key="cart_item.id">
             <td style="text-align: center;">{{ cart_item.cake_name }}</td>
             <td style="text-align: center;" >
-              <button @click="decrement(cart_item)" class="px-2 mr-2" style=" background-color: transparent !important; border: none; ">
+              <button @click="decrement(cart_item)" @change="total_price" class="px-2 mr-2" style=" background-color: transparent !important; border: none; ">
                   <b-icon icon="dash" aria-hidden="true"></b-icon> 
               </button> 
               {{ cart_item.quantity }}
-              <button @click="increment(cart_item)" class="px-2 ml-2" style=" background-color: transparent !important; border: none; ">
+              <button @click="increment(cart_item)" @change="total_price" class="px-2 ml-2" style=" background-color: transparent !important; border: none; ">
                   <b-icon icon="plus" aria-hidden="true"></b-icon> 
               </button> 
             </td>
@@ -28,25 +28,36 @@
                 <td colspan="2" style="text-align: right;">Subtotal:</td>
                 <td style="text-align: center;">{{ total }}</td>
             </tr>
+            <tr v-if="selectedOption == 'delivery'">
+                <td class="border-top-0" colspan="2" style="text-align: right;">Deli:</td>
+                <td class="border-top-0" style="text-align: center;"> {{ deli_price }} </td>
+            </tr>
             <tr>
                 <td class="border-top-0" colspan="2" style="text-align: right;">Tax:</td>
                 <td class="border-top-0" style="text-align: center;">{{ tax }}</td>
             </tr>
             <tr>
                 <td colspan="2" style="text-align: right; font-weight: bold;">Total:</td>
-                <td style="font-weight: bold; text-align: center;">{{ total + tax }}</td>
+                <td style="font-weight: bold; text-align: center;">{{ total + tax + deli_price}}</td>
             </tr>
         </tbody>
         </table>
-        <label>
-            <input type="radio" v-model="selectedOption" value="cash_on_deli"> Cash On Delivery
-        </label>
-        <label>
-            <input type="radio" v-model="selectedOption" value="pick_up"> Pick Up
-        </label>
+        <div class="d-flex align-items-center justify-content-center">
+          <label>
+            <input type="radio" v-model="selectedOption" value="delivery" @change="delivery_price"> Cash On Delivery
+          </label>
+          <label class="ml-2">
+            <input type="radio" v-model="selectedOption" value="pick_up" @change="delivery_price"> Pick Up
+          </label>
+        </div>
         <div class="clear-fix">
           <button class="float-left mt-2 btn btn-outline-dark" @click="back">Back</button>
-          <button class="float-right mt-2 btn btn-outline-info" @click="save_order">Confirm Order</button>
+          <button class="float-right mt-2 btn btn-outline-info"
+              @click="save_order"
+              :disabled="isConfirmButtonDisabled"
+              :class="{ 'disabled': isConfirmButtonDisabled }">
+               Confirm Order
+          </button>
         </div>
       </div>
     </div>
@@ -59,7 +70,9 @@ export default{
            cart_items: [],
            total: 0,
            tax: 0,
-           selectedOption: ''
+           selectedOption: '',
+           amount: 0,
+           deli_price: 0
         }
     },
     methods: {
@@ -74,35 +87,71 @@ export default{
         const percentage = 5;
         this.tax = (this.total * percentage) / 100;
       },
+      delivery_price() {
+        console.log(this.selectedOption,"sel")
+        this.deli_price = this.selectedOption === 'delivery' ? 2000 : 0;
+      },
       save_order(){
-        this.cart_items.map(cart_item =>{
-          cart_item['option'] = this.selectedOption
-        })
-        this.$axios.post(`/user_details`,this.cart_items)
-        .then(response => {
-          alert("Your Order Have Saved Successfully")
-          localStorage.clear();
-          this.$router.push('/')
-          
-        })
-        .catch(error => {
-          this.$notify({
-            title: 'Fail',
-            text: 'Something went wrong. Please try again',
-            type: 'error'
-          });
-          this.errors = error.response.data.error
-          this.errorMessage = true
-        })
+        if( this.amount >= this.total)
+        {
+          const postData = {
+            cart_items: this.cart_items,
+            selectedOption: this.selectedOption,
+            total_price: this.total+this.tax+this.deli_price
+          };
+          this.$axios.post(`/orders`, postData)
+          .then(response => {
+            alert("Your Order Have Saved Successfully")
+            localStorage.clear();
+            this.$router.push('/')
+            
+          })
+          .catch(error => {
+            this.$notify({
+              title: 'Fail',
+              text: 'Something went wrong. Please try again',
+              type: 'error'
+            });
+            this.errors = error.response.data.error
+            this.errorMessage = true
+          })
+        }else
+        {
+          alert("Your balance is insufficient! Plase Top Up");
+        }
       },
       back(){
         this.$router.back();
-      }
+      },
+      increment(cart_item){
+        cart_item.quantity ++;
+        this.updateLocalStorage();
+        this.total_price(),
+        this.tax_price()
+      },
+      decrement(cart_item){
+        if (cart_item.quantity != 1){
+        cart_item.quantity --;
+        this.updateLocalStorage();
+        this.total_price(),
+        this.tax_price()
+        }
+      },
+      updateLocalStorage() {
+        localStorage.setItem('cart_items', JSON.stringify(this.cart_items));
+      },
     },
     mounted() {
-      this.cart_items = this.$route.params['cartitems']
+      this.cart_items = JSON.parse(localStorage.getItem('cart_items')) || [];
+      this.amount = this.$route.params['amount']
       this.total_price(),
-      this.tax_price()
+      this.tax_price(),
+      this.delivery_price
+    },
+    computed: {
+      isConfirmButtonDisabled() {
+        return !this.selectedOption; // Returns true if no radio option is selected
+      }
     }
 }
 </script>
